requirements.txt
requests
python-magic
pefile
rich